fix(contracts): resolve finalization artifacts
Use Foundry's signature-specific simulation artifact and share resilient Cast transaction-hash parsing across release boundaries.
